refactor(runtime): split wasm_fs into a wasm package (abi, bridge, fs, host, runtime)
One module per job: abi.py holds the preview1 wire constants and
record packers (distinct numbering from POSIX, so nothing is reusable
from errno/os), bridge.py the sync dispatch hop, fs.py the path
router, host.py the preview1 host functions, runtime.py the shared
WasmRuntime. Tests mirror the split; test_host.py guards the ABI
table against the method signatures.
